Add support for ssize_t in TensorShape
PiperOrigin-RevId: 305053288

+++ tensorflow/python/lib/core/py_seq_tensor.cc
@@ -72,7 +72,7 @@ bool IsPyFloat(PyObject* obj) {
 
 struct ConverterState {
   // The inferred tensor shape.
-  gtl::InlinedVector<int64, 4> inferred_shape;
+  gtl::InlinedVector<ssize_t, 4> inferred_shape;
 
   // The inferred tensor data type.
   DataType inferred_dtype;
@@ -320,7 +320,7 @@ struct ConverterTraits<int64> {
   }
 
   static AbstractTensorInterface* CreateTensor(
-      TFE_Context* ctx, absl::Span<const int64> dim_sizes) {
+      TFE_Context* ctx, absl::Span<const ssize_t> dim_sizes) {
     return ctx->context->CreateInt64Tensor(dim_sizes);
   }
 
@@ -360,7 +360,7 @@ struct ConverterTraits<uint64> {
   }
 
   static AbstractTensorInterface* CreateTensor(
-      TFE_Context* ctx, absl::Span<const int64> dim_sizes) {
+      TFE_Context* ctx, absl::Span<const ssize_t> dim_sizes) {
     return ctx->context->CreateUint64Tensor(dim_sizes);
   }
 
@@ -397,7 +397,7 @@ struct ConverterTraits<int32> {
   }
 
   static AbstractTensorInterface* CreateTensor(
-      TFE_Context* ctx, absl::Span<const int64> dim_sizes) {
+      TFE_Context* ctx, absl::Span<const ssize_t> dim_sizes) {
     return ctx->context->CreateInt32Tensor(dim_sizes);
   }
 
@@ -504,7 +504,7 @@ struct ConverterTraits<float> {
   }
 
   static AbstractTensorInterface* CreateTensor(
-      TFE_Context* ctx, absl::Span<const int64> dim_sizes) {
+      TFE_Context* ctx, absl::Span<const ssize_t> dim_sizes) {
     return ctx->context->CreateFloatTensor(dim_sizes);
   }
 
@@ -520,7 +520,7 @@ struct ConverterTraits<double> {
   }
 
   static AbstractTensorInterface* CreateTensor(
-      TFE_Context* ctx, absl::Span<const int64> dim_sizes) {
+      TFE_Context* ctx, absl::Span<const ssize_t> dim_sizes) {
     return ctx->context->CreateDoubleTensor(dim_sizes);
   }
 
@@ -540,7 +540,7 @@ struct ConverterTraits<Eigen::half> {
   }
 
   static AbstractTensorInterface* CreateTensor(
-      TFE_Context* ctx, absl::Span<const int64> dim_sizes) {
+      TFE_Context* ctx, absl::Span<const ssize_t> dim_sizes) {
     return ctx->context->CreateHalfTensor(dim_sizes);
   }
 
@@ -561,7 +561,7 @@ struct ConverterTraits<tstring> {
   }
 
   static AbstractTensorInterface* CreateTensor(
-      TFE_Context* ctx, absl::Span<const int64> dim_sizes) {
+      TFE_Context* ctx, absl::Span<const ssize_t> dim_sizes) {
     return ctx->context->CreateStringTensor(dim_sizes);
   }
 
@@ -628,7 +628,7 @@ struct ConverterTraits<complex128> {
   }
 
   static AbstractTensorInterface* CreateTensor(
-      TFE_Context* ctx, absl::Span<const int64> dim_sizes) {
+      TFE_Context* ctx, absl::Span<const ssize_t> dim_sizes) {
     return ctx->context->CreateComplex128Tensor(dim_sizes);
   }
 
@@ -656,7 +656,7 @@ struct ConverterTraits<bool> {
   }
 
   static AbstractTensorInterface* CreateTensor(
-      TFE_Context* ctx, absl::Span<const int64> dim_sizes) {
+      TFE_Context* ctx, absl::Span<const ssize_t> dim_sizes) {
     return ctx->context->CreateBoolTensor(dim_sizes);
   }
